backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-cors-headers>=4.6.0",
    "django-cotton>=1.5.1",
    "django-debug-toolbar==4.4.6",
    "django-extensions==3.2.3",
    "django-phonenumber-field==8.0.0",
    "django-seed==0.3.1",
    "django-unfold==0.40.0",
    "django==5.1.2",
    "djangorestframework-simplejwt>=5.4.0",
    "djangorestframework==3.15.2",
    "faker==30.6.0",
    "googletrans==3.0.0",
    "gtts==2.5.3",
    "opencv-python==4.10.0.84",
    "phonenumbers==8.13.47",
    "pillow==11.0.0",
    "python-dotenv==1.0.1",
    "whitenoise==6.7.0",
]
